requirements.txt
gradio
tensorflow
mlflow
dagshub
pandas
numpy
pillow
cv2
mtcnn